+++ src/components/User/SystemAssets/Issuance/components/IssuanceHistory.vue
@@ -53,9 +53,9 @@
 
             <span
               class="app-list__cell app-list__cell--right app-list__cell--important"
-              :title="`${localize(item.details.issuanceCreate.amount)} ${item.details.issuanceCreate.asset}`"
+              :title="`${localize(item.details.createIssuance.amount)} ${item.details.createIssuance.asset}`"
             >
-              {{ localize(item.details.issuanceCreate.amount) }} {{item.details.issuanceCreate.asset }}
+              {{ localize(item.details.createIssuance.amount) }} {{item.details.createIssuance.asset }}
             </span>
 
             <span
@@ -67,9 +67,9 @@
 
             <span
               class="app-list__cell app-list__cell--right "
-              :title="item.details.issuanceCreate.receiver"
+              :title="item.details.createIssuance.receiver"
             >
-              <email-getter :balance-id="item.details.issuanceCreate.receiver" />
+              <email-getter :balance-id="item.details.createIssuance.receiver" />
             </span>
 
             <span